requirement.txt
Mysql-connector-python
fastapi[all]
